requirements.txt
argon2-cffi==21.3.0
fastapi==0.85.0
furs-fiscal==1.0.0
loguru==0.6.0
psycopg2==2.9.4
pycryptodome==3.15.0
python-multipart==0.0.5
pytz==2022.4
SQLAlchemy==1.4.41
uvicorn==0.18.3
